Restore: stake real coins and settle them through the ledger (#30 never reached main)

+++ BetSocial/src/main/java/com/example/World/Bets/BetDTO.java
@@ -23,8 +23,8 @@ public record BetDTO(
      @NonNull
      Boolean profit_mode,
      @NonNull
-     Float max_amount,
+     Long max_amount,
      @NonNull
-     Float min_amount
+     Long min_amount
     ) {
 }
